+++ docs/CHANGELOG.md
@@ -1,3 +1,13 @@
 # CHANGELOG
+
+## v0.1
 - 2024-11-16: Created the project.
 - 2024-11-16: Compile and run the "Hello, World!" program.
+- 2024-11-16: Added the `CMake` build script.
+- 2024-11-16: Added the batch scripts to build/configure/run on Windows.
+- 2024-11-17: Created the pre-complied header file `base.pch.hpp`.
+- 2024-11-17: Added the base aliases for fixed size integer types to the `pch`.
+- 2024-11-17: Added `spdlog-v1.14` as dependency to the project.
+- 2024-11-17: Added logging utilities to the `pch`, implementation uses `spdlog`.
+
+## v0.2